cmpe273-comm-models-lab/streaming-kafka/inventory_consumer/requirements.txt
Flask==2.3.0
kafka-python==2.0.2
python-json-logger==2.0.7
